First version completed but not yet fully tested.

+++ src/main/java/com/data/orderbook/domain/model/OrderBookCandle.java
@@ -3,7 +3,13 @@ package com.data.orderbook.domain.model;
 import java.math.BigDecimal;
 
 public record OrderBookCandle(
-        String pair, Long timestamp, BigDecimal open, BigDecimal high, BigDecimal low, BigDecimal close, Long ticks) {
+        String pair,
+        Long timestamp,
+        BigDecimal open,
+        BigDecimal high,
+        BigDecimal low,
+        BigDecimal close,
+        Integer ticks) {
     // - Timestamp; timestamp at the start of the minute (usually in epoch)
     // - Open; mid price at the start of the minute
     // - High; highest mid price during the minute
